requirements.txt
nlpcube==0.1.0.8
ufal.udpipe==1.2.0.3
numpy==1.16.0
scipy==1.2.0
scikit-learn==0.17
filelock
pytest
flask
flask-restful
